add delete fcm device on delete secure user

+++ first_api/views.py
@@ -1304,15 +1304,14 @@ class SecureGuardViewSet(viewsets.ModelViewSet):
             secure.secure_username = None
             secure.save()
             
-
-            result = serializer.data
-            
-            # print("usernamePk")
-            # print(usernamePk)
-
-            username = models.UserProfile.objects.get(pk =usernamePk)
-            username.delete()
-            username.save
+            isUsernameExist = models.UserProfile.objects.filter(pk =usernamePk).exists()
+            if(isUsernameExist==True):
+                username = models.UserProfile.objects.get(pk =usernamePk)
+                ### delete fcm device
+                fcmDevice = models.CustomFCMDevice.objects.filter(user=username).delete()
+                ### delete username profile
+                username.delete()
+                username.save
 
             return Response(serializer.data)
